ButtonPattern excluido e InputPassword criado
