+++ Semana 12 - Revisão/Projeto - LabEddit/labeddit/src/pages/Register/RegisterForm.js
@@ -2,12 +2,12 @@ import { useState } from 'react'
 import { useHistory } from 'react-router-dom'
 import useForm from '../../hooks/useForm'
 import { ViewIcon, ViewOffIcon } from '@chakra-ui/icons'
-import { Button, ContainerButton } from '../../styles/ButtonFormStyles'
-import { InputGroup, InputRightElement } from '@chakra-ui/react'
+import { ButtonPattern, ContainerButton } from '../../styles/ButtonFormStyles'
+import { InputGroup, InputRightElement, Button } from '@chakra-ui/react'
 import { InputPattern } from "../../components/InputPattern"
-import { ButtonPattern } from '../../components/ButtonPattern'
 import signup from '../../requests/signup'
 import { token } from '../../APIConfig/token'
+import InputPassword from '../../components/InputPassword'
 
 function Registerform() {
   const [show, setShow] = useState(false)
@@ -53,31 +53,13 @@ function Registerform() {
       value={form.email}
       onChange={onChange}
     />
-
-    <InputGroup>
-      <InputPattern
-        placeholder="Senha"
-        name={'password'}
-        value={form.password}
-        onChange={onChange}
-        type={show ? "text" : "password"}
-      />
-      <InputRightElement>
-        <ButtonPattern
-          marginBottom={'-.1rem'}
-          onClick={handleClick}
-          background={'rgba(255, 255, 255, 0)'}
-          color={'#000'}
-          bgHover={'white'}
-          text={show ? <ViewIcon /> : <ViewOffIcon />}
-          boxShadow={'none'}
-          bgActive={'white'}
-        />
-      </InputRightElement>
-    </InputGroup>
-
+    <InputPassword
+      value={form.password}
+      onChange={onChange}
+    />
+    
     <ContainerButton>
-      <Button type={'submit'}>Cadastrar</Button>
+      <ButtonPattern type={'submit'}>Cadastrar</ButtonPattern>
     </ContainerButton>
   </form>
 }
