+++ Semana 12 - Revisão/Projeto - LabEddit/labeddit/src/styles/ButtonFormStyles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-export const Button = styled.button `
+export const ButtonPattern = styled.button `
   border-radius: 20px;
   background: #050533;
   color: white;
